feat: added service layer
